Make XD Server location configurable for accessing the aggregate-counters.
The base URL of the xd server can now be configured via the xdServerBaseUrl property.

java -jar target/smartgrid-frontend-1.0-SNAPSHOT.jar --xdServerBaseUrl=http://localhost:9393

The configured location is checked for reachability on startup.
An error message will be logged to the console in when the location could not be reached.

2014-09-04 11:45:29.305 ERROR 25924 --- [           main] f.t.AggregateCounterTimeSeriesRepository : Could not connect to server: http://xxx:9393 Error: UnknownHostException: xxx